+++ src/Lucid/Relations/BaseRelation.js
@@ -43,6 +43,8 @@ const methodsList = [
   'delete',
   'update',
   'first',
+  'last',
+  'firstOrFail',
   'fetch',
   'toSQL',
   'toString'
@@ -91,7 +93,7 @@ class BaseRelation {
      * @return {void}
      */
     this._eagerLoadFn = function (query, fk, values) {
-      query.whereIn(fk, values)
+      query.whereIn(`${this.RelatedModel.table}.${fk}`, values)
     }
 
     /**