+++ src/Migration/index.js
@@ -30,6 +30,7 @@ class Migration {
     this.db = Database
     this._migrationsTable = Config.get('database.migrationsTable', 'adonis_schema')
     this._lockTable = `${this._migrationsTable}_lock`
+    this.isKeepAliveEnabled = false
   }
 
   /**
@@ -164,7 +165,7 @@ class Migration {
    * @private
    */
   _addForBatch (name, batch) {
-    return this.db.table(this._migrationsTable).insert({name, batch})
+    return this.db.table(this._migrationsTable).insert({ name, batch })
   }
 
   /**
@@ -226,8 +227,8 @@ class Migration {
    */
   async _getDiff (names, direction = 'up', batch) {
     const schemas = direction === 'down'
-    ? await this._getAfterBatch(batch)
-    : await this.db.table(this._migrationsTable).pluck('name')
+      ? await this._getAfterBatch(batch)
+      : await this.db.table(this._migrationsTable).pluck('name')
 
     return direction === 'down' ? _.reverse(_.intersection(names, schemas)) : _.difference(names, schemas)
   }
@@ -301,7 +302,23 @@ class Migration {
    */
   async _cleanup () {
     await this._removeLock()
-    this.db.close()
+
+    if (!this.isKeepAliveEnabled) {
+      this.db.close()
+    }
+  }
+
+  /**
+   * Enable or disable keepAlive, which prevents the database connection from being closed.
+   *
+   * @method keepAlive
+   *
+   * @param {boolean}enabled
+   *
+   * @return {void}
+   */
+  keepAlive (enabled = true) {
+    this.isKeepAliveEnabled = enabled
   }
 
   /**
@@ -456,7 +473,10 @@ class Migration {
       .table(this._migrationsTable)
       .orderBy('name')
 
-    this.db.close()
+    if (!this.isKeepAliveEnabled) {
+      this.db.close()
+    }
+
     return _.map(schemas, (schema, name) => {
       const migration = _.find(migrated, (mig) => mig.name === name)
       return {